+++ klog/app/cli/create.go
@@ -35,10 +35,6 @@ func (opt *Create) Run(ctx app.Context) app.Error {
 		[]reconciling.Creator{
 			reconciling.NewReconcilerForNewRecord(date, opt.DateFormat(ctx.Config()), additionalData),
 		},
-
-		func(reconciler *reconciling.Reconciler) error {
-			return nil
-		},
 	)
 }